feat(autosync): move music code to survey
