+++ package.json
@@ -79,6 +79,7 @@
     "keytime": "^0.1.0",
     "once": "^1.3.1",
     "pixi.js": "^2.2.3",
-    "ramda": "^0.9.1"
+    "ramda": "^0.9.1",
+    "vue": "^0.11.4"
   }
 }
